added scpi streaming API , fixed bug with timestamp timer and streaming timer

+++ firmware/src/HAL/TimerApi/TimerApi.c
@@ -75,7 +75,7 @@ void TimerApi_Stop(uint8_t index) {
     }
 }
 
-void TimerApi_PeriodSet(uint8_t index, uint16_t period) {
+void TimerApi_PeriodSet(uint8_t index, uint32_t period) {
     switch (index) {
         case 2:
             TMR2_PeriodSet(period);
@@ -94,7 +94,7 @@ void TimerApi_PeriodSet(uint8_t index, uint16_t period) {
     }
 }
 
-uint16_t TimerApi_PeriodGet(uint8_t index) {
+uint32_t TimerApi_PeriodGet(uint8_t index) {
     uint16_t ret = 0;
     switch (index) {
         case 2:
@@ -115,7 +115,7 @@ uint16_t TimerApi_PeriodGet(uint8_t index) {
     return ret;
 }
 
-uint16_t TimerApi_CounterGet(uint8_t index) {
+uint32_t TimerApi_CounterGet(uint8_t index) {
     uint16_t ret = 0;
     switch (index) {
         case 2:
